added id to the token payload
